Prioritise team numbers over most criteria met
This allows users to prioritise equal team numbers over the most
accurate representation of the criteria. It was developed in response
to team builders with complex criteria having some teams that varied
from two to five members.

In our testing, from very simple to very complex criteria, prioritise
numbers gave no worse results than the current algorithm, although it's
possible that in some circumstances it would do so. For this reason,
prioritise even numbers has been made the default, whilst the option to
use the old algorithm is still available.