Moodle 2.2 compatibility
Moodle 2.2 disallows optional_param taking an array, and disallows
arrays of arrays in parameters. This changes the teams array of arrays
submitted via POST when creating groups to two arrays, teams and
teamnames, with teams being an array of comma-separated user IDs rather
than an array of arrays of user IDs.